Add controller node to cluster
Currently, we don't setup kubelet and kube-proxy on the controller node,
i.e. the node is not part of the cluster nodes and no pods can be
deployed on it.

Change that in order to enable kubedee users to test workloads which
expect controller nodes to be part of the cluster, for example to test
daemon sets that are supposed to (also) run on a controller.

Also, rename '${name}' -> '${cluster_name}' throughout the whole code
base for clarity.