Add `--kubernetes-version` option
Currently, if users want to install a cluster based on an upstream
version of Kubernetes, they have to download the binaries themself (e.g.
by using the `download-k8s-binaries` helper script) and then specify
the location with `--bin-dir`.

Instead, add a new option `--kubernetes-version` that takes precedence
over `--bin-dir` and do that work for the user.