Store k8s binaries in cache dir
This is a follow up to

   b16327e7c2 Add `--kubernetes-version` option

Switch to storing the Kubernetes binaries in the kubedee cache dir.
While this means that the binaries will have to be downloaded again
after updating kubedee (as it is the case for all other cached files),
it is more future-proof: we know that we start from a clean state
after updating kubedee and thus are free to make changes to the code.