lib: don't run `apt-get upgrade`
kubedee's base image gets regenerated fairly regularly (with every
kubedee version change) and is based on LXD's `ubuntu:18.04` image,
which gets updated regularly as well. So the packages should be mostly
up-to-date (or up-to-date "enough) and we can save a bit of setup time
with avoiding the upgrade.

We can also save a bit of disk space which sometimes is an issue, for
example in a CI environment.

Also, remove apt's cache before creating the container image.